+++ src/main/java/innotutor/innotutor_backend/controller/user/StudentsController.java
@@ -1,9 +1,9 @@
 package innotutor.innotutor_backend.controller.user;
 
-import innotutor.innotutor_backend.dto.enrollment.tutor.RespondedTutorsListInfoDTO;
+import innotutor.innotutor_backend.dto.enrollment.student.RequestedStudentsListInfoDTO;
 import innotutor.innotutor_backend.security.CustomPrincipal;
 import innotutor.innotutor_backend.service.CardEnrollService;
-import innotutor.innotutor_backend.service.TutorsService;
+import innotutor.innotutor_backend.service.StudentsService;
 import innotutor.innotutor_backend.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -12,39 +12,34 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(value = "/my-tutors", produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(value = "/my-students", produces = MediaType.APPLICATION_JSON_VALUE)
 @CrossOrigin(origins = "*", allowedHeaders = "*", methods = {RequestMethod.GET, RequestMethod.PUT, RequestMethod.DELETE})
-public class MyTutorsController {
-    private final transient TutorsService tutorsService;
+public class StudentsController {
+    private final transient StudentsService studentsService;
     private final transient CardEnrollService cardEnrollService;
     private final transient UserService userService;
 
-    public MyTutorsController(final TutorsService tutorsService,
-                              final CardEnrollService cardEnrollService,
-                              final UserService userService) {
-        this.tutorsService = tutorsService;
+    public StudentsController(final StudentsService studentsService, final CardEnrollService cardEnrollService, final UserService userService) {
+        this.studentsService = studentsService;
         this.cardEnrollService = cardEnrollService;
         this.userService = userService;
     }
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<RespondedTutorsListInfoDTO> getUserTutorsList(
-            @AuthenticationPrincipal final CustomPrincipal user) {
-        final RespondedTutorsListInfoDTO tutors
-                = tutorsService.getUserTutorsListFullInfo(userService.getUserId(user));
-        return tutors == null
+    public ResponseEntity<RequestedStudentsListInfoDTO> getStudentsList(@AuthenticationPrincipal final CustomPrincipal user) {
+        final RequestedStudentsListInfoDTO students = studentsService.getUserStudentsListFullInfo(userService.getUserId(user));
+        return students == null
                 ? new ResponseEntity<>(HttpStatus.NOT_FOUND)
-                : new ResponseEntity<>(tutors, HttpStatus.OK);
+                : new ResponseEntity<>(students, HttpStatus.OK);
     }
 
     @PutMapping(value = "accept/{enrollmentId}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> acceptTutor(@PathVariable final Long enrollmentId,
-                                         @AuthenticationPrincipal final CustomPrincipal user) {
+    public ResponseEntity<?> acceptStudent(@PathVariable final Long enrollmentId, @AuthenticationPrincipal final CustomPrincipal user) {
         ResponseEntity<?> response;
         if (enrollmentId == null) {
             response = new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         } else {
-            response = cardEnrollService.acceptTutor(userService.getUserId(user), enrollmentId)
+            response = cardEnrollService.acceptStudent(userService.getUserId(user), enrollmentId)
                     ? new ResponseEntity<>(HttpStatus.OK)
                     : new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
@@ -52,13 +47,12 @@ public class MyTutorsController {
     }
 
     @DeleteMapping(value = "remove/{enrollmentId}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> removeTutor(@PathVariable final Long enrollmentId,
-                                         @AuthenticationPrincipal final CustomPrincipal user) {
+    public ResponseEntity<?> removeStudent(@PathVariable final Long enrollmentId, @AuthenticationPrincipal final CustomPrincipal user) {
         ResponseEntity<?> response;
         if (enrollmentId == null) {
             response = new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         } else {
-            response = cardEnrollService.removeTutor(userService.getUserId(user), enrollmentId)
+            response = cardEnrollService.removeStudent(userService.getUserId(user), enrollmentId)
                     ? new ResponseEntity<>(HttpStatus.OK)
                     : new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
