+++ src/main/java/innotutor/innotutor_backend/controller/user/TutorsController.java
@@ -1,9 +1,9 @@
 package innotutor.innotutor_backend.controller.user;
 
-import innotutor.innotutor_backend.dto.enrollment.student.RequestedStudentsListInfoDTO;
+import innotutor.innotutor_backend.dto.enrollment.tutor.RespondedTutorsListInfoDTO;
 import innotutor.innotutor_backend.security.CustomPrincipal;
 import innotutor.innotutor_backend.service.CardEnrollService;
-import innotutor.innotutor_backend.service.StudentsService;
+import innotutor.innotutor_backend.service.TutorsService;
 import innotutor.innotutor_backend.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -12,34 +12,39 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(value = "/my-students", produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(value = "/my-tutors", produces = MediaType.APPLICATION_JSON_VALUE)
 @CrossOrigin(origins = "*", allowedHeaders = "*", methods = {RequestMethod.GET, RequestMethod.PUT, RequestMethod.DELETE})
-public class MyStudentsController {
-    private final transient StudentsService studentsService;
+public class TutorsController {
+    private final transient TutorsService tutorsService;
     private final transient CardEnrollService cardEnrollService;
     private final transient UserService userService;
 
-    public MyStudentsController(final StudentsService studentsService, final CardEnrollService cardEnrollService, final UserService userService) {
-        this.studentsService = studentsService;
+    public TutorsController(final TutorsService tutorsService,
+                            final CardEnrollService cardEnrollService,
+                            final UserService userService) {
+        this.tutorsService = tutorsService;
         this.cardEnrollService = cardEnrollService;
         this.userService = userService;
     }
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<RequestedStudentsListInfoDTO> getUserStudentsList(@AuthenticationPrincipal final CustomPrincipal user) {
-        final RequestedStudentsListInfoDTO students = studentsService.getUserStudentsListFullInfo(userService.getUserId(user));
-        return students == null
+    public ResponseEntity<RespondedTutorsListInfoDTO> getTutorsList(
+            @AuthenticationPrincipal final CustomPrincipal user) {
+        final RespondedTutorsListInfoDTO tutors
+                = tutorsService.getUserTutorsListFullInfo(userService.getUserId(user));
+        return tutors == null
                 ? new ResponseEntity<>(HttpStatus.NOT_FOUND)
-                : new ResponseEntity<>(students, HttpStatus.OK);
+                : new ResponseEntity<>(tutors, HttpStatus.OK);
     }
 
     @PutMapping(value = "accept/{enrollmentId}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> acceptStudent(@PathVariable final Long enrollmentId, @AuthenticationPrincipal final CustomPrincipal user) {
+    public ResponseEntity<?> acceptTutor(@PathVariable final Long enrollmentId,
+                                         @AuthenticationPrincipal final CustomPrincipal user) {
         ResponseEntity<?> response;
         if (enrollmentId == null) {
             response = new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         } else {
-            response = cardEnrollService.acceptStudent(userService.getUserId(user), enrollmentId)
+            response = cardEnrollService.acceptTutor(userService.getUserId(user), enrollmentId)
                     ? new ResponseEntity<>(HttpStatus.OK)
                     : new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
@@ -47,12 +52,13 @@ public class MyStudentsController {
     }
 
     @DeleteMapping(value = "remove/{enrollmentId}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> removeStudent(@PathVariable final Long enrollmentId, @AuthenticationPrincipal final CustomPrincipal user) {
+    public ResponseEntity<?> removeTutor(@PathVariable final Long enrollmentId,
+                                         @AuthenticationPrincipal final CustomPrincipal user) {
         ResponseEntity<?> response;
         if (enrollmentId == null) {
             response = new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         } else {
-            response = cardEnrollService.removeStudent(userService.getUserId(user), enrollmentId)
+            response = cardEnrollService.removeTutor(userService.getUserId(user), enrollmentId)
                     ? new ResponseEntity<>(HttpStatus.OK)
                     : new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
